+++ src/it/java/org/greencodeinitiative/creedengo/python/integration/tests/GCIRulesIT.java
@@ -347,6 +347,22 @@ class GCIRulesIT extends GCIRulesBase {
             7, 11, 16, 21, 45, 52, 60
         };
         checkIssuesForFile(filePath, ruleId, ruleMsg, startLines, endLines, SEVERITY, TYPE, EFFORT_5MIN);
+
+    @Test 
+    void testGCI107(){
+
+        String filePath = "src/avoidIterativeMatrixOperations.py";
+        String ruleId = "creedengo-python:GCI107";
+        String ruleMsg = "Avoid iterative matrix operations, use numpy dot or outer function instead";
+        int[] startLines = new int[]{
+            8, 20, 36, 46, 75, 83, 91, 106, 115
+        };
+        int[] endLines = new int[]{
+            8, 20, 36, 46, 75, 83, 91, 106, 115
+        };
+
+        checkIssuesForFile(filePath, ruleId, ruleMsg, startLines, endLines, SEVERITY, TYPE, EFFORT_1h);
+      
     }