+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- [#68](https://github.com/green-code-initiative/creedengo-python/pull/68) Add rule GCI107 avoidIterativeMatrixOperations
+
 ### Changed
 
 ### Deleted